[GAIA Compiler] integrate ci tests under different storages
