+++ interactive_engine/compiler/src/main/java/com/alibaba/graphscope/gremlin/integration/suite/pattern/IrPatternTestSuite.java
@@ -14,31 +14,31 @@
  * limitations under the License.
  */
 
-package com.alibaba.graphscope.integration.ldbc;
+package com.alibaba.graphscope.gremlin.integration.suite.pattern;
 
 import org.apache.tinkerpop.gremlin.AbstractGremlinSuite;
 import org.apache.tinkerpop.gremlin.process.traversal.TraversalEngine;
 import org.junit.runners.model.InitializationError;
 import org.junit.runners.model.RunnerBuilder;
 
-public class IrLdbcTestSuite extends AbstractGremlinSuite {
+public class IrPatternTestSuite extends AbstractGremlinSuite {
 
     private static final Class<?>[] allTests =
             new Class<?>[] {
-                LdbcQueryTest.Traversals.class,
+                PatternQueryTest.Traversals.class,
             };
 
     private static final Class<?>[] testsToEnforce =
             new Class<?>[] {
-                LdbcQueryTest.Traversals.class,
+                PatternQueryTest.Traversals.class,
             };
 
-    public IrLdbcTestSuite(final Class<?> klass, final RunnerBuilder builder)
+    public IrPatternTestSuite(final Class<?> klass, final RunnerBuilder builder)
             throws InitializationError {
         super(klass, builder, allTests, testsToEnforce, false, TraversalEngine.Type.STANDARD);
     }
 
-    public IrLdbcTestSuite(
+    public IrPatternTestSuite(
             final Class<?> klass, final RunnerBuilder builder, final Class<?>[] testsToExecute)
             throws InitializationError {
         super(klass, builder, testsToExecute, testsToEnforce, true, TraversalEngine.Type.STANDARD);
