+++ interactive_engine/compiler/src/test/java/com/alibaba/graphscope/gremlin/PathExpandStepTest.java
@@ -18,7 +18,7 @@ package com.alibaba.graphscope.gremlin;
 
 import com.alibaba.graphscope.common.intermediate.operator.PathExpandOp;
 import com.alibaba.graphscope.common.jna.type.*;
-import com.alibaba.graphscope.gremlin.antlr4.__;
+import com.alibaba.graphscope.gremlin.integration.suite.utils.__;
 import com.alibaba.graphscope.gremlin.plugin.traversal.IrCustomizedTraversalSource;
 import com.alibaba.graphscope.gremlin.transform.StepTransformFactory;
