+++ interactive_engine/tests/pom.xml
@@ -19,6 +19,10 @@
       <artifactId>sdk-common</artifactId>
     </dependency>
     <dependency>
+      <groupId>com.alibaba.graphscope</groupId>
+      <artifactId>compiler</artifactId>
+    </dependency>
+    <dependency>
       <groupId>org.apache.commons</groupId>
       <artifactId>commons-lang3</artifactId>
     </dependency>
