+++ interactive_engine/tests/src/test/java/com/alibaba/graphscope/function/test/gremlin/GremlinStandardTest.java
@@ -15,8 +15,8 @@
  */
 package com.alibaba.graphscope.function.test.gremlin;
 
-import com.alibaba.graphscope.function.test.RemoteTestGraph;
 import com.alibaba.graphscope.function.test.RemoteTestGraphProvider;
+import com.alibaba.graphscope.gremlin.integration.graph.RemoteTestGraph;
 
 import org.apache.tinkerpop.gremlin.GraphProviderClass;
 import org.junit.runner.RunWith;
